hw4/requirements.txt
numpy
sklearn.svm
gen
word2vec
nltk
cv2
os
PIL
matplotlib.cm
matplotlib.pyplot
math